requirements.txt
numpy
pyaudio
soundfile
pvporcupine
speech_recognition
vk_api